User member role in tests that use users
